Updates to user role and journal entries
Updates to the controllers for user roles specifically for
enabling/disabling roles and journal entries and the logic with saving
entries.

+++ src/main/java/com/journal/repositories/UserRoleRepository.java
@@ -23,7 +23,7 @@ public interface UserRoleRepository extends JpaRepository <UserRole, Integer> {
     @Query(value = "delete from UserRole u where u.username = ?1 and u.role = ?2")
     void deleteUserRoleByUsernameAndRole(String username, String role);
 
-    UserRole findByUserRoleId(int userRoleId);
+    UserRole findByUserRoleId(long userRoleId);
 
     List<UserRole> findByUsername(String username);
